+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,43 +1,10 @@
 package com.techcourse.service;
 
-import com.interface21.transaction.support.TransactionTemplate;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
 import java.util.Optional;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class UserService {
-
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final TransactionTemplate transactionTemplate;
-
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao, final TransactionTemplate transactionTemplate) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionTemplate = transactionTemplate;
-    }
-
-    public Optional<User> findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        final var user = findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("해당 id의 유저가 존재하지 않습니다" + id));
-
-        transactionTemplate.execute(() -> {
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-        });
-    }
+public interface UserService {
+    Optional<User> findById(final long id);
+    void save(final User user);
+    void changePassword(final long id, final String newPassword, final String createdBy);
 }
